camera.py thread snapshot.py
