+++ custom_components/valetudo_vacuum_camera/utils/colors_man.py
@@ -136,102 +136,113 @@ color_array = [
 ]
 
 
-def add_alpha_to_rgb(alpha_channels, rgb_colors):
-    """
-    Add alpha channel to RGB colors using corresponding alpha channels.
+class ColorsManagment:
+    """Class to manage the colors.
+    Imports and updates the colors from the user configuration."""
 
-    Args:
-        alpha_channels (List[Optional[float]]): List of alpha channel values (0.0-255.0).
-        rgb_colors (List[Tuple[int, int, int]]): List of RGB colors.
+    def __init__(self, shared_var):
+        self.shared_var = shared_var
 
-    Returns:
-        List[Tuple[int, int, int, int]]: List of RGBA colors with alpha channel added.
-    """
-    if len(alpha_channels) != len(rgb_colors):
-        _LOGGER.error("Input lists must have the same length.")
-        return []
+    @staticmethod
+    def add_alpha_to_rgb(alpha_channels, rgb_colors):
+        """
+        Add alpha channel to RGB colors using corresponding alpha channels.
 
-    result = []
-    for alpha, rgb in zip(alpha_channels, rgb_colors):
+        Args:
+            alpha_channels (List[Optional[float]]): List of alpha channel values (0.0-255.0).
+            rgb_colors (List[Tuple[int, int, int]]): List of RGB colors.
+
+        Returns:
+            List[Tuple[int, int, int, int]]: List of RGBA colors with alpha channel added.
+        """
+        if len(alpha_channels) != len(rgb_colors):
+            _LOGGER.error("Input lists must have the same length.")
+            return []
+
+        result = []
+        for alpha, rgb in zip(alpha_channels, rgb_colors):
+            try:
+                alpha_int = int(alpha)
+                if alpha_int < 0:
+                    alpha_int = 0
+                elif alpha_int > 255:
+                    alpha_int = 255
+
+                if rgb is None:
+                    result.append((0, 0, 0, alpha_int))
+                else:
+                    result.append((rgb[0], rgb[1], rgb[2], alpha_int))
+            except (ValueError, TypeError):
+                result.append(None)
+
+        return result
+
+    def set_initial_colours(self, device_info: dict) -> None:
+        """Set the initial colours for the map."""
         try:
-            alpha_int = int(alpha)
-            if alpha_int < 0:
-                alpha_int = 0
-            elif alpha_int > 255:
-                alpha_int = 255
-
-            if rgb is None:
-                result.append((0, 0, 0, alpha_int))
-            else:
-                result.append((rgb[0], rgb[1], rgb[2], alpha_int))
-        except (ValueError, TypeError):
-            result.append(None)
-
-    return result
-
-
-def set_initial_colours(device_info: dict, shared_var) -> None:
-    """Set the initial colours for the map."""
-    try:
-        user_colors = [
-            device_info.get(COLOR_WALL),
-            device_info.get(COLOR_ZONE_CLEAN),
-            device_info.get(COLOR_ROBOT),
-            device_info.get(COLOR_BACKGROUND),
-            device_info.get(COLOR_MOVE),
-            device_info.get(COLOR_CHARGER),
-            device_info.get(COLOR_NO_GO),
-            device_info.get(COLOR_GO_TO),
-            device_info.get(COLOR_TEXT),
-        ]
-        user_alpha = [
-            device_info.get(ALPHA_WALL),
-            device_info.get(ALPHA_ZONE_CLEAN),
-            device_info.get(ALPHA_ROBOT),
-            device_info.get(ALPHA_BACKGROUND),
-            device_info.get(ALPHA_MOVE),
-            device_info.get(ALPHA_CHARGER),
-            device_info.get(ALPHA_NO_GO),
-            device_info.get(ALPHA_GO_TO),
-            device_info.get(ALPHA_TEXT),
-        ]
-        rooms_colors = [
-            device_info.get(COLOR_ROOM_0),
-            device_info.get(COLOR_ROOM_1),
-            device_info.get(COLOR_ROOM_2),
-            device_info.get(COLOR_ROOM_3),
-            device_info.get(COLOR_ROOM_4),
-            device_info.get(COLOR_ROOM_5),
-            device_info.get(COLOR_ROOM_6),
-            device_info.get(COLOR_ROOM_7),
-            device_info.get(COLOR_ROOM_8),
-            device_info.get(COLOR_ROOM_9),
-            device_info.get(COLOR_ROOM_10),
-            device_info.get(COLOR_ROOM_11),
-            device_info.get(COLOR_ROOM_12),
-            device_info.get(COLOR_ROOM_13),
-            device_info.get(COLOR_ROOM_14),
-            device_info.get(COLOR_ROOM_15),
-        ]
-        rooms_alpha = [
-            device_info.get(ALPHA_ROOM_0),
-            device_info.get(ALPHA_ROOM_1),
-            device_info.get(ALPHA_ROOM_2),
-            device_info.get(ALPHA_ROOM_3),
-            device_info.get(ALPHA_ROOM_4),
-            device_info.get(ALPHA_ROOM_5),
-            device_info.get(ALPHA_ROOM_6),
-            device_info.get(ALPHA_ROOM_7),
-            device_info.get(ALPHA_ROOM_8),
-            device_info.get(ALPHA_ROOM_9),
-            device_info.get(ALPHA_ROOM_10),
-            device_info.get(ALPHA_ROOM_11),
-            device_info.get(ALPHA_ROOM_12),
-            device_info.get(ALPHA_ROOM_13),
-            device_info.get(ALPHA_ROOM_14),
-            device_info.get(ALPHA_ROOM_15),
-        ]
-        shared_var.update_user_colors(add_alpha_to_rgb(user_alpha, user_colors))
-        shared_var.update_rooms_colors(add_alpha_to_rgb(rooms_alpha, rooms_colors))
-    except (ValueError, IndexError, UnboundLocalError) as e:
-        _LOGGER.error("Error while populating colors: %s", e)
+            user_colors = [
+                device_info.get(COLOR_WALL),
+                device_info.get(COLOR_ZONE_CLEAN),
+                device_info.get(COLOR_ROBOT),
+                device_info.get(COLOR_BACKGROUND),
+                device_info.get(COLOR_MOVE),
+                device_info.get(COLOR_CHARGER),
+                device_info.get(COLOR_NO_GO),
+                device_info.get(COLOR_GO_TO),
+                device_info.get(COLOR_TEXT),
+            ]
+            user_alpha = [
+                device_info.get(ALPHA_WALL),
+                device_info.get(ALPHA_ZONE_CLEAN),
+                device_info.get(ALPHA_ROBOT),
+                device_info.get(ALPHA_BACKGROUND),
+                device_info.get(ALPHA_MOVE),
+                device_info.get(ALPHA_CHARGER),
+                device_info.get(ALPHA_NO_GO),
+                device_info.get(ALPHA_GO_TO),
+                device_info.get(ALPHA_TEXT),
+            ]
+            rooms_colors = [
+                device_info.get(COLOR_ROOM_0),
+                device_info.get(COLOR_ROOM_1),
+                device_info.get(COLOR_ROOM_2),
+                device_info.get(COLOR_ROOM_3),
+                device_info.get(COLOR_ROOM_4),
+                device_info.get(COLOR_ROOM_5),
+                device_info.get(COLOR_ROOM_6),
+                device_info.get(COLOR_ROOM_7),
+                device_info.get(COLOR_ROOM_8),
+                device_info.get(COLOR_ROOM_9),
+                device_info.get(COLOR_ROOM_10),
+                device_info.get(COLOR_ROOM_11),
+                device_info.get(COLOR_ROOM_12),
+                device_info.get(COLOR_ROOM_13),
+                device_info.get(COLOR_ROOM_14),
+                device_info.get(COLOR_ROOM_15),
+            ]
+            rooms_alpha = [
+                device_info.get(ALPHA_ROOM_0),
+                device_info.get(ALPHA_ROOM_1),
+                device_info.get(ALPHA_ROOM_2),
+                device_info.get(ALPHA_ROOM_3),
+                device_info.get(ALPHA_ROOM_4),
+                device_info.get(ALPHA_ROOM_5),
+                device_info.get(ALPHA_ROOM_6),
+                device_info.get(ALPHA_ROOM_7),
+                device_info.get(ALPHA_ROOM_8),
+                device_info.get(ALPHA_ROOM_9),
+                device_info.get(ALPHA_ROOM_10),
+                device_info.get(ALPHA_ROOM_11),
+                device_info.get(ALPHA_ROOM_12),
+                device_info.get(ALPHA_ROOM_13),
+                device_info.get(ALPHA_ROOM_14),
+                device_info.get(ALPHA_ROOM_15),
+            ]
+            self.shared_var.update_user_colors(
+                self.add_alpha_to_rgb(user_alpha, user_colors)
+            )
+            self.shared_var.update_rooms_colors(
+                self.add_alpha_to_rgb(rooms_alpha, rooms_colors)
+            )
+        except (ValueError, IndexError, UnboundLocalError) as e:
+            _LOGGER.error("Error while populating colors: %s", e)
